backend/requirements.txt
Django>=5.0
psycopg2-binary
djangorestframework>=3.15
djangorestframework-simplejwt>=5.4
django-cors-headers>=4.4
python-jose[cryptography]
requests